+++ skorch/net.py
@@ -1422,17 +1422,17 @@ class NeuralNet:
             self.initialize_callbacks()
             self._set_params_callback(**cb_params)
 
-        if any(key.startswith('criterion') for key in special_params):
+        if any('criterion' in key.split('__', 1)[0] for key in special_params):
             self.initialize_criterion()
 
         module_triggers_optimizer_reinit = False
-        if any(key.startswith('module') for key in special_params):
+        if any('module' in key.split('__', 1)[0] for key in special_params):
             self.initialize_module()
             module_triggers_optimizer_reinit = True
 
         optimizer_changed = (
-            any(key.startswith('optimizer') for key in special_params) or
-            'lr' in normal_params
+            any('optimizer' in key.split('__', 1)[0] for key in special_params)
+            or 'lr' in normal_params
         )
         if module_triggers_optimizer_reinit or optimizer_changed:
             # Model selectors such as GridSearchCV will set the
@@ -1523,6 +1523,92 @@ class NeuralNet:
         state.pop('__cuda_dependent_attributes__')
 
         self.__dict__.update(state)
+
+    def _register_attribute(
+            self,
+            name,
+            prefixes=True,
+            cuda_dependent_attributes=True,
+    ):
+        """Add attribute name to prefixes_ and
+        cuda_dependent_attributes_.
+
+        Takes care of not mutating the lists.
+
+        Parameters
+        ----------
+        prefixes : bool (default=True)
+          Whether to add to prefixes_.
+
+        cuda_dependent_attributes : bool (default=True)
+          Whether to add to cuda_dependent_attributes_.
+
+        """
+        # copy the lists to avoid mutation
+        if prefixes:
+            self.prefixes_ = self.prefixes_[:] + [name]
+
+        if cuda_dependent_attributes:
+            self.cuda_dependent_attributes_ = (
+                self.cuda_dependent_attributes_[:] + [name + '_'])
+
+    def _unregister_attribute(
+            self,
+            name,
+            prefixes=True,
+            cuda_dependent_attributes=True,
+    ):
+        """Remove attribute name from prefixes_ and
+        cuda_dependent_attributes_.
+
+        Takes care of not mutating the lists.
+
+        Parameters
+        ----------
+        prefixes : bool (default=True)
+          Whether to remove from prefixes_.
+
+        cuda_dependent_attributes : bool (default=True)
+          Whether to remove from cuda_dependent_attributes_.
+
+        """
+        # copy the lists to avoid mutation
+        if prefixes:
+            self.prefixes_ = [p for p in self.prefixes_[:] if p != name]
+
+        if cuda_dependent_attributes:
+            self.cuda_dependent_attributes_ = [
+                a for a in self.cuda_dependent_attributes_ if a != name + '_']
+
+    def __setattr__(self, name, attr):
+        """Set an attribute on the net
+
+        When a custom net with additional torch modules or optimizers
+        is created, those attributes are added to ``prefixes_`` and
+        ``cuda_dependent_attributes_`` automatically.
+
+        """
+        # if it's a known attribute or not a torch module/optimizer
+        # instance or class, just setattr as usual
+        is_known = (name.endswith('_') or (name in self.prefixes_))
+        is_torch_mod_or_opt = (
+            isinstance(attr, (torch.optim.Optimizer, torch.nn.Module))
+            or (
+                isinstance(attr, type)
+                and issubclass(attr, (torch.optim.Optimizer, torch.nn.Module))
+            )
+        )
+        if is_known or not is_torch_mod_or_opt:
+            super().__setattr__(name, attr)
+            return
+
+        self._register_attribute(name)
+        super().__setattr__(name, attr)
+
+    def __delattr__(self, name):
+        # take extra precautions to undo the changes made in __setattr__
+        self._unregister_attribute(name)
+        super().__delattr__(name)
 
     def save_params(
             self, f_params=None, f_optimizer=None, f_history=None):